produce granules for each download url
Previously, it was assumed there was a single downloadable granule for
each UMM metadata item. In the case multiple downloadable granules were
found it incorrect concatenated values separated by a NL.

This changes it so each download url or download direct url results in a
distinct granule that can be downloaded.

Generally, we would expect a single granule, but this should support
cases where provider metadata has multiple granules or provides granule
metadata that is wonky.